renamed itemsidebar to shopsidebar for clarifcation
