+++ client/src/assets/css/shop.css
@@ -74,3 +74,7 @@
   .filters-header {
     margin-top: 10px;
   }
+
+  .filter-items {
+    margin-top: 10px;
+  }
