+++ client/src/components/index.js
@@ -2,7 +2,7 @@ export { default as Header } from './Header'
 export { default as ProfileInfo } from './ProfileInfo'
 export { default as Buttons } from './Buttons'
 export { default as Stats } from './Stats'
-export { default as ItemSidebar } from './ItemSidebar'
+export { default as ShopSidebar } from './ShopSidebar'
 export { default as SearchItem } from './SearchItem'
 export { default as Footer } from './Footer'
 export { default as Founders } from './Founders'
